making health api public.
