+++ pkg/handlers/health.go
@@ -1,0 +1,35 @@
+/****************************************************************************
+ * Copyright 2020, Optimizely, Inc. and contributors                        *
+ *                                                                          *
+ * Licensed under the Apache License, Version 2.0 (the "License");          *
+ * you may not use this file except in compliance with the License.         *
+ * You may obtain a copy of the License at                                  *
+ *                                                                          *
+ *    http://www.apache.org/licenses/LICENSE-2.0                            *
+ *                                                                          *
+ * Unless required by applicable law or agreed to in writing, software      *
+ * distributed under the License is distributed on an "AS IS" BASIS,        *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
+ * See the License for the specific language governing permissions and      *
+ * limitations under the License.                                           *
+ ***************************************************************************/
+
+// Package handlers //
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/go-chi/render"
+)
+
+// HealthBody is holding info about health checks
+type HealthBody struct {
+	Status  string   `json:"status,omitempty"`
+	Reasons []string `json:"reasons,omitempty"`
+}
+
+// Health displays health status
+func Health(w http.ResponseWriter, r *http.Request) {
+	render.JSON(w, r, HealthBody{Status: "ok"})
+}
